refactoring and separate 2 classes into separate headers
as file was growing a lot
